tests/multibot/requirements.txt
requests>=2.31.0
pyyaml>=6.0
pytest>=7.4.0
pytest-timeout>=2.2.0
